+++ src/Problems/BBOB/f_weierstrass.hpp
@@ -93,7 +93,7 @@ public:
   }
 
   double internal_evaluate(const std::vector<double> &x) {
-    int n = x.size();
+    size_t n = x.size();
     size_t i, j;
 
     std::vector<double> result(1);
